pyproject.toml
[project]
name = "core-services"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "django-cors-headers>=4.3.1",
]
